third_party/blink/web_tests/platform/generic/storage/websql/null-characters-expected.txt

Test that embedded NULL (U+0000) characters don't truncate strings.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is "foo\u0000bar"
PASS successfullyParsed is true

TEST COMPLETE
